algaworks-especialista-jpa/algaworks-ecommerce/src/main/java/br/com/jkavdev/algaworks/ejpa/ecommerce/model/Cliente.java
package br.com.jkavdev.algaworks.ejpa.ecommerce.model;

import javax.persistence.Entity;
import javax.persistence.Id;
import java.util.Objects;

@Entity
public class Cliente {

    @Id
    private Integer id;

    private String nome;

    // precisamos do construtor padrao, caso queiramos trazer essa entidade do banco de dados
    protected Cliente() {
    }

    public Cliente(String nome) {
        this.nome = nome;
    }

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    @Override
    public boolean equals(final Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        final Cliente cliente = (Cliente) o;
        return getId().equals(cliente.getId());
    }

    @Override
    public int hashCode() {
        return Objects.hash(getId());
    }
}
